requirements.txt
# Core dependencies
Pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0

# AI/ML dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
diffusers>=0.21.0
accelerate>=0.20.0

# Image processing
scikit-image>=0.21.0
imageio>=2.31.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
click>=8.1.0
tqdm>=4.65.0
requests>=2.31.0
